Updating FSMC driver (SRAM part)
