+++ os/hal/include/fsmc/sram.h
@@ -43,32 +43,32 @@
  * @brief   SRAM driver enable switch.
  * @details If set to @p TRUE the support for SRAM1 is included.
  */
-#if !defined(STM32_SRAM_USE_FSMC_SRAM1) || defined(__DOXYGEN__)
-#define STM32_SRAM_USE_FSMC_SRAM1                  FALSE
+#if !defined(STM32_FSMC_USE_SRAM1) || defined(__DOXYGEN__)
+#define STM32_FSMC_USE_SRAM1                  FALSE
 #endif
 
 /**
  * @brief   SRAM driver enable switch.
  * @details If set to @p TRUE the support for SRAM2 is included.
  */
-#if !defined(STM32_SRAM_USE_FSMC_SRAM2) || defined(__DOXYGEN__)
-#define STM32_SRAM_USE_FSMC_SRAM2                  FALSE
+#if !defined(STM32_FSMC_USE_SRAM2) || defined(__DOXYGEN__)
+#define STM32_FSMC_USE_SRAM2                  FALSE
 #endif
 
 /**
  * @brief   SRAM driver enable switch.
  * @details If set to @p TRUE the support for SRAM3 is included.
  */
-#if !defined(STM32_SRAM_USE_FSMC_SRAM3) || defined(__DOXYGEN__)
-#define STM32_SRAM_USE_FSMC_SRAM3                  FALSE
+#if !defined(STM32_FSMC_USE_SRAM3) || defined(__DOXYGEN__)
+#define STM32_FSMC_USE_SRAM3                  FALSE
 #endif
 
 /**
  * @brief   SRAM driver enable switch.
  * @details If set to @p TRUE the support for SRAM4 is included.
  */
-#if !defined(STM32_SRAM_USE_FSMC_SRAM4) || defined(__DOXYGEN__)
-#define STM32_SRAM_USE_FSMC_SRAM4                  FALSE
+#if !defined(STM32_FSMC_USE_SRAM4) || defined(__DOXYGEN__)
+#define STM32_FSMC_USE_SRAM4                  FALSE
 #endif
 
 /** @} */
@@ -77,13 +77,13 @@
 /* Derived constants and error checks.                                       */
 /*===========================================================================*/
 
-#if !STM32_SRAM_USE_FSMC_SRAM1 && !STM32_SRAM_USE_FSMC_SRAM2 && \
-    !STM32_SRAM_USE_FSMC_SRAM3 && !STM32_SRAM_USE_FSMC_SRAM4
+#if !STM32_FSMC_USE_SRAM1 && !STM32_FSMC_USE_SRAM2 && \
+    !STM32_FSMC_USE_SRAM3 && !STM32_FSMC_USE_SRAM4
 #error "SRAM driver activated but no SRAM peripheral assigned"
 #endif
 
-#if (STM32_SRAM_USE_FSMC_SRAM1 || STM32_SRAM_USE_FSMC_SRAM2 || \
-    STM32_SRAM_USE_FSMC_SRAM3 || STM32_SRAM_USE_FSMC_SRAM4) && !STM32_HAS_FSMC
+#if (STM32_FSMC_USE_SRAM1 || STM32_FSMC_USE_SRAM2 || \
+    STM32_FSMC_USE_SRAM3 || STM32_FSMC_USE_SRAM4) && !STM32_HAS_FSMC
 #error "FSMC not present in the selected device"
 #endif
 
@@ -137,19 +137,19 @@ struct SRAMDriver {
 /* External declarations.                                                    */
 /*===========================================================================*/
 
-#if STM32_SRAM_USE_FSMC_SRAM1 && !defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM1 && !defined(__DOXYGEN__)
 extern SRAMDriver SRAMD1;
 #endif
 
-#if STM32_SRAM_USE_FSMC_SRAM2 && !defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM2 && !defined(__DOXYGEN__)
 extern SRAMDriver SRAMD2;
 #endif
 
-#if STM32_SRAM_USE_FSMC_SRAM3 && !defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM3 && !defined(__DOXYGEN__)
 extern SRAMDriver SRAMD3;
 #endif
 
-#if STM32_SRAM_USE_FSMC_SRAM4 && !defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM4 && !defined(__DOXYGEN__)
 extern SRAMDriver SRAMD4;
 #endif
