+++ os/hal/ports/STM32/LLD/FSMCv1/hal_fsmc_sram_lld.c
@@ -22,9 +22,10 @@
  * @{
  */
 #include "hal.h"
-#include "hal_fsmc_sram.h"
 
-#if (STM32_USE_FSMC_SRAM == TRUE) || defined(__DOXYGEN__)
+#if (HAL_USE_FSMC_SRAM == TRUE) || defined(__DOXYGEN__)
+
+#include "hal_fsmc_sram_lld.h"
 
 /*===========================================================================*/
 /* Driver local definitions.                                                 */
@@ -36,28 +37,28 @@
 /**
  * @brief   SRAM1 driver identifier.
  */
-#if STM32_SRAM_USE_FSMC_SRAM1 || defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM1 || defined(__DOXYGEN__)
 SRAMDriver SRAMD1;
 #endif
 
 /**
  * @brief   SRAM2 driver identifier.
  */
-#if STM32_SRAM_USE_FSMC_SRAM2 || defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM2 || defined(__DOXYGEN__)
 SRAMDriver SRAMD2;
 #endif
 
 /**
  * @brief   SRAM3 driver identifier.
  */
-#if STM32_SRAM_USE_FSMC_SRAM3 || defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM3 || defined(__DOXYGEN__)
 SRAMDriver SRAMD3;
 #endif
 
 /**
  * @brief   SRAM4 driver identifier.
  */
-#if STM32_SRAM_USE_FSMC_SRAM4 || defined(__DOXYGEN__)
+#if STM32_FSMC_USE_SRAM4 || defined(__DOXYGEN__)
 SRAMDriver SRAMD4;
 #endif
 
@@ -82,36 +83,6 @@ SRAMDriver SRAMD4;
 /*===========================================================================*/
 
 /**
- * @brief   Low level SRAM driver initialization.
- *
- * @notapi
- */
-void fsmcSramInit(void) {
-
-  fsmc_init();
-
-#if STM32_SRAM_USE_FSMC_SRAM1
-  SRAMD1.sram = FSMCD1.sram1;
-  SRAMD1.state = SRAM_STOP;
-#endif /* STM32_SRAM_USE_FSMC_SRAM1 */
-
-#if STM32_SRAM_USE_FSMC_SRAM2
-  SRAMD2.sram = FSMCD1.sram2;
-  SRAMD2.state = SRAM_STOP;
-#endif /* STM32_SRAM_USE_FSMC_SRAM2 */
-
-#if STM32_SRAM_USE_FSMC_SRAM3
-  SRAMD3.sram = FSMCD1.sram3;
-  SRAMD3.state = SRAM_STOP;
-#endif /* STM32_SRAM_USE_FSMC_SRAM3 */
-
-#if STM32_SRAM_USE_FSMC_SRAM4
-  SRAMD4.sram = FSMCD1.sram4;
-  SRAMD4.state = SRAM_STOP;
-#endif /* STM32_SRAM_USE_FSMC_SRAM4 */
-}
-
-/**
  * @brief   Configures and activates the SRAM peripheral.
  *
  * @param[in] sramp         pointer to the @p SRAMDriver object
@@ -119,20 +90,11 @@ void fsmcSramInit(void) {
  *
  * @notapi
  */
-void fsmcSramStart(SRAMDriver *sramp, const SRAMConfig *cfgp) {
+void lld_sram_start(SRAMDriver *sramp, const SRAMConfig *cfgp) {
 
-  if (FSMCD1.state == FSMC_STOP)
-    fsmc_start(&FSMCD1);
-
-  osalDbgAssert((sramp->state == SRAM_STOP) || (sramp->state == SRAM_READY),
-              "invalid state");
-
-  if (sramp->state == SRAM_STOP) {
     sramp->sram->BTR  = cfgp->btr;
     sramp->sram->BWTR = cfgp->bwtr;
     sramp->sram->BCR  = cfgp->bcr | FSMC_BCR_MBKEN;
-    sramp->state = SRAM_READY;
-  }
 }
 
 /**
@@ -142,9 +104,8 @@ void fsmcSramStart(SRAMDriver *sramp, const SRAMConfig *cfgp) {
  *
  * @notapi
  */
-void fsmcSramStop(SRAMDriver *sramp) {
+void lld_sram_stop(SRAMDriver *sramp) {
 
-  if (sramp->state == SRAM_READY) {
     uint32_t mask = FSMC_BCR_MBKEN;
 #if (defined(STM32F427xx) || defined(STM32F437xx) || \
      defined(STM32F429xx) || defined(STM32F439xx) || \
@@ -155,11 +116,9 @@ void fsmcSramStop(SRAMDriver *sramp) {
     mask |= FSMC_BCR_CCLKEN;
 #endif
     sramp->sram->BCR &= ~mask;
-    sramp->state = SRAM_STOP;
-  }
 }
 
-#endif /* STM32_USE_FSMC_SRAM */
+#endif /* STM32_FSMC_USE_SRAM */
 
 /** @} */
